Add support for Debian 9.
This patch adds support for bulding Shadow on Debian 9 (stretch). On
Debian 9 the debug symbols of ld.so have to be found within the
`/usr/lib/debug/.build-id/...` directory tree. The checksum that is used
to find the debug symbols in the build-id tree can be found using the
`readelf -n file` command.