hack gdb performance with shadow linkmap This commit adds a shadow linkmap to use internally in elf-loader, leaving the linkmap used by the glibc ABI untouched after dynamic linking (i.e., dynamically loaded objects will *not* be added to glibc's linkmap). This has the effect of making it feasible to run shadow experiments within gdb, or to attach gdb to a long-running shadow experiment, since it won't load the debugging symbols for any plugins (an operation that took ~linear time in the number of objects in the glibc ABI linkmap, per object in the linkmap, so quadratic time in practice).
If a user needs to debug a plugin in gdb, they can call the elf-loader function
vdl_linkmap_abi_update() from gdb, which will load all objects in the shadow
linkmap into the glibc ABI linkmap at once, at which point gdb will load those
debug symbols. I don't know if this is any faster than the amount of time it
took originally to just let gdb run, but this at least lets the user choose
based on the size of the experiment if they want to do this. The next commit
will add the ability to choose individual instances to load.